FieldType works now

+++ WasaBii-unity-project/Assets/Geometry/dump.cs
@@ -7,7 +7,7 @@ using WasaBii.Geometry.Shared;
 namespace BII.WasaBii.Geometry
 {
     
-    [GeometryHelper(areFieldsIndependent: true, fieldType: (int)FieldType.Length, hasMagnitude: true, hasDirection: true)]
+    [GeometryHelper(areFieldsIndependent: true, fieldType: FieldType.Length, hasMagnitude: true, hasDirection: true)]
     public partial struct GlobalOffset3 {
         public readonly Length X, Y, Z;
         public Length test() => this.Magnitude;
